+++ celt/mdct_short_blocks_test.go
@@ -2,7 +2,7 @@ package celt
 
 import "testing"
 
-func mdctShortBlocksCoreLegacy(samples []float64, overlap, shortBlocks, shortSize int, output, blockCoeffs []float64, blockMDCT func(block, coeffs []float64)) {
+func mdctShortBlocksCoreLegacy(samples []float32, overlap, shortBlocks, shortSize int, output, blockCoeffs []float32, blockMDCT func(block, coeffs []float32)) {
 	for b := 0; b < shortBlocks; b++ {
 		start := b * shortSize
 		end := start + shortSize + overlap
@@ -19,17 +19,17 @@ func mdctShortBlocksCoreLegacy(samples []float64, overlap, shortBlocks, shortSiz
 	}
 }
 
-func mdctShortBlocksStub(block, coeffs []float64) {
-	base := float64(len(block))
+func mdctShortBlocksStub(block, coeffs []float32) {
+	base := float32(len(block))
 	for i := range coeffs {
-		coeffs[i] = base + float64(i)*0.25
+		coeffs[i] = base + float32(i)*0.25
 	}
 }
 
 func TestMDCTShortBlocksCoreMatchesLegacy(t *testing.T) {
-	samples := make([]float64, 1080)
+	samples := make([]float32, 1080)
 	for i := range samples {
-		samples[i] = float64((i%31)-15) * 0.125
+		samples[i] = float32((i%31)-15) * 0.125
 	}
 	const (
 		overlap     = 120
@@ -37,10 +37,10 @@ func TestMDCTShortBlocksCoreMatchesLegacy(t *testing.T) {
 		shortSize   = 120
 	)
 	outputLen := shortBlocks * shortSize
-	want := make([]float64, outputLen)
-	got := make([]float64, outputLen)
-	blockCoeffsWant := make([]float64, shortSize)
-	blockCoeffsGot := make([]float64, shortSize)
+	want := make([]float32, outputLen)
+	got := make([]float32, outputLen)
+	blockCoeffsWant := make([]float32, shortSize)
+	blockCoeffsGot := make([]float32, shortSize)
 
 	mdctShortBlocksCoreLegacy(samples, overlap, shortBlocks, shortSize, want, blockCoeffsWant, mdctShortBlocksStub)
 	mdctShortBlocksCore(samples, overlap, shortBlocks, shortSize, got, blockCoeffsGot, mdctShortBlocksStub)
@@ -53,17 +53,17 @@ func TestMDCTShortBlocksCoreMatchesLegacy(t *testing.T) {
 }
 
 func benchmarkMDCTShortBlocksCore(b *testing.B, legacy bool) {
-	samples := make([]float64, 1080)
+	samples := make([]float32, 1080)
 	for i := range samples {
-		samples[i] = float64((i%31)-15) * 0.125
+		samples[i] = float32((i%31)-15) * 0.125
 	}
 	const (
 		overlap     = 120
 		shortBlocks = 8
 		shortSize   = 120
 	)
-	output := make([]float64, shortBlocks*shortSize)
-	blockCoeffs := make([]float64, shortSize)
+	output := make([]float32, shortBlocks*shortSize)
+	blockCoeffs := make([]float32, shortSize)
 
 	b.ReportAllocs()
 	b.ResetTimer()
